Fixed issues
* Optimized fragment animations
* Migrated all data retrievals to the repository
* Changed back stack handling
* Optimized communication between parent activity and fragments
* Optimized data classes

+++ app/src/main/java/com/example/hris/repository/HrisRepository.kt
@@ -1,9 +1,10 @@
 package com.example.hris.repository
 
 import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
 import com.example.hris.data.HrisDao
-import com.example.hris.model.TimeLogs
-import com.example.hris.model.User
+import com.example.hris.model.*
+import com.example.hris.network.HrisApi
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
 import javax.inject.Inject
@@ -12,11 +13,13 @@ import javax.inject.Singleton
 @Singleton
 class HrisRepository @Inject constructor(
     private val hrisDao: HrisDao
-){
+) {
 
     val profileData: LiveData<User> = hrisDao.getProfile()
 
     val timeLogs: LiveData<List<TimeLogs>> = hrisDao.getTimeLogs()
+    val timeLogsResponse = MutableLiveData<TimeLogsModel>()
+    val addTimeLogsResponse = MutableLiveData<AddTimeLogsResponseModel>()
 
     suspend fun refreshProfile(user: User) {
         withContext(Dispatchers.IO) {
@@ -25,10 +28,26 @@ class HrisRepository @Inject constructor(
         }
     }
 
-    suspend fun refreshTimeLogs(timeLogs: List<TimeLogs>) {
+    suspend fun refreshTimeLogs() {
         withContext(Dispatchers.IO) {
-            hrisDao.deleteTimeLogs()
-            hrisDao.insertTimeLogs(timeLogs)
+            val call = HrisApi.retrofitService.getTimeLogs(
+                profileData.value!!.userID
+            )
+            if (call.status == "0") {
+                hrisDao.deleteTimeLogs()
+                hrisDao.insertTimeLogs(call.timeLogs!!)
+            }
+            timeLogsResponse.postValue(call)
+        }
+    }
+
+    suspend fun addTimeLogs(type: String) {
+        withContext(Dispatchers.IO) {
+            val call = HrisApi.retrofitService.addTimeLogs(
+                profileData.value!!.userID,
+                type
+            )
+            addTimeLogsResponse.postValue(call)
         }
     }
 }
